+++ src/main/java/thb/fbi/leguan/instructions/IBranchCode.java
@@ -6,5 +6,5 @@ import thb.fbi.leguan.simulation.PCRegister;
  * Interface for definining and executing code of unconditional branch instructions
  */
 public interface IBranchCode {
-    public abstract void simulate(int br_address, PCRegister pc);
+    public abstract void simulate(long br_address, PCRegister pc);
 }
