+++ src/main/java/thb/fbi/leguan/instructions/IConditionalBranchCode.java
@@ -7,5 +7,5 @@ import thb.fbi.leguan.simulation.IntegerRegister;
  * Interface for definining and executing code of conditional branch instructions
  */
 public interface IConditionalBranchCode {
-    public abstract void simulate(int cond_br_address, IntegerRegister Rt, PCRegister pc);
+    public abstract void simulate(long cond_br_address, IntegerRegister Rt, PCRegister pc);
 }
