+++ src/main/java/thb/fbi/pipeline_visualizer/pipeline/MEM_WB_Pipeline.java
@@ -9,7 +9,7 @@ public class MEM_WB_Pipeline implements Serializable{
     /** WriteBack stage */
     public WB  WB;
     /** result from the ALU */
-    public int ALU_result;
+    public long ALU_result;
     /** data loaded from memory */
     public int memoryData;
     /** destination register index */
@@ -19,6 +19,6 @@ public class MEM_WB_Pipeline implements Serializable{
     /** original line of code/instruction */
     public String codeString;
     /** PC of the next instruction */
-    public int PC; // for branch prediction needed
+    public long PC; // for branch prediction needed
     
 }
